requirements.txt
requests
pandas
pyarrow
python-dotenv
apscheduler
streamlit>=1.32.0
plotly
sqlalchemy
